Add files via upload
Fixed the bug in the loop for best_cluster_selection. Previously looped from 0 to number of block in cluster (could be from say 0 to 164). This is incorrect. Should loop through number of clusters (never more than say, 10). This has been fixed by putting in a reference to the number of clusters as the upper limit.

Also, configured the state of the file such that it is set up for direct calculation of SBS4 hcal detection efficiency -- as long as you are using jboyd's parsed files and my included DB of cuts, limits, etc. 

Last checked this runs and calculates SBS4 detection efficiency at about 98.2%